Fix TimeSeriesForestRegressor (#777)
* Fix TimeSeriesForestRegressor

* Add TSF regressor to docs
