+++ docs/source/api_reference/regression.rst
@@ -3,8 +3,7 @@
 sktime.regression: Time series regression
 =========================================
 
-The :mod:`sktime.regression` module contains algorithms and composition
-tools for time series regression.
+The :mod:`sktime.regression` module contains algorithms and composition tools for time series regression.
 
 .. automodule:: sktime.regression
     :no-members:
@@ -14,6 +13,17 @@ Composition
 -----------
 
 .. currentmodule:: sktime.regression.compose
+
+.. autosummary::
+    :toctree: modules/auto_generated/
+    :template: class.rst
+
+    ComposableTimeSeriesForestRegressor
+
+Interval-based
+--------------
+
+.. currentmodule:: sktime.regression.interval_based
 
 .. autosummary::
     :toctree: modules/auto_generated/
