+++ sktime/regression/all/__init__.py
@@ -2,9 +2,15 @@
 # -*- coding: utf-8 -*-
 
 __author__ = ["Markus Löning"]
-__all__ = ["np", "pd", "ComposableTimeSeriesForestRegressor"]
+__all__ = [
+    "np",
+    "pd",
+    "ComposableTimeSeriesForestRegressor",
+    "TimeSeriesForestRegressor",
+]
 
 import numpy as np
 import pandas as pd
 
 from sktime.regression.compose import ComposableTimeSeriesForestRegressor
+from sktime.regression.interval_based import TimeSeriesForestRegressor
