+++ sktime/regression/interval_based/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3 -u
+# -*- coding: utf-8 -*-
+
+__author__ = ["Markus Löning"]
+__all__ = ["TimeSeriesForestRegressor"]
+
+from sktime.regression.interval_based._tsf import TimeSeriesForestRegressor
