+++ sktime/series_as_features/base/estimators/interval_based/__init__.py
@@ -1,2 +1,8 @@
 # -*- coding: utf-8 -*-
 # Empty for now
+
+__all__ = ["BaseTimeSeriesForest"]
+
+from sktime.series_as_features.base.estimators.interval_based._tsf import (
+    BaseTimeSeriesForest,
+)
